+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## Unreleased
 - Minor text/doc changes
+- Added `2021-01` API version to enum. [#117](https://github.com/shopify/shopify-node-api/pull/117)
+
 ## [1.0.0]
 
 - Initial public release